Add trainer class
